+++ packages/core/src/ledger/index.ts
@@ -1,2 +1,3 @@
 export * from './ledger.js'
 export * from './deposits.js'
+export * from './helius.js'